+++ lib/Search.tsx
@@ -6,11 +6,32 @@ import SearchIcon from './components/icons/SearchIcon';
 import Key from './components/key/Key';
 import SearchModal from './components/search-modal/SearchModal';
 import classes from './Search.module.css';
-import type { SearchProps } from './Search.d';
-
 import './global.css';
 
-function Search({
+export type EnhancedSearchConfig = {
+  projectId: string;
+  accessToken: string;
+}
+
+export type DocSearch = {
+  apiKey: string;
+  host: string;
+  collection: string;
+}
+
+export type Config = {
+  enhancedSearch: EnhancedSearchConfig;
+  docSearch?: DocSearch;
+}
+
+export type SearchProps = {
+  config: Config;
+  placeholder?: string;
+  size?: 'small' | 'middle' | 'large';
+  shape?: 'square' | 'round';
+}
+
+export default function Search ({
   config,
   placeholder = 'Search',
   size = 'middle',
@@ -56,5 +77,3 @@ function Search({
     </>
   )
 }
-
-export default Search;