+++ lib/components/icons/LinkIcon.tsx
@@ -1,4 +1,4 @@
-function LinkIcon() {
+export default function LinkIcon () {
   return (
     <svg stroke="currentColor" fill="currentColor" strokeWidth="0" viewBox="0 0 24 24" height="1em" width="1em"
       xmlns="http://www.w3.org/2000/svg">
@@ -7,5 +7,3 @@ function LinkIcon() {
     </svg>
   );
 }
-
-export default LinkIcon;