+++ lib/components/key/Key.d.ts
@@ -1,5 +1,0 @@
-export type KeyProps = {
-  children: React.ReactNode;
-  className?: string;
-  size?: 'small' | 'middle' | 'large';
-}