+++ lib/components/key/Key.tsx
@@ -1,8 +1,13 @@
 import { classNames } from '../../helpers/styles';
 import classes from './Key.module.css';
-import type { KeyProps } from './Key.d';
 
-function Key({ children, className, size = 'middle', ...props }: KeyProps) {
+type KeyProps = {
+  children: React.ReactNode;
+  className?: string;
+  size?: 'small' | 'middle' | 'large';
+}
+
+export default function Key ({ children, className, size = 'middle', ...props }: KeyProps) {
   return (
     <div
       className={
@@ -18,5 +23,3 @@ function Key({ children, className, size = 'middle', ...props }: KeyProps) {
     </div>
   )
 }
-
-export default Key;