+++ lib/components/search-modal/SearchModal.d.ts
@@ -1,7 +1,0 @@
-import type { Config } from '../../Search.d';
-
-export type SearchModalProps = {
-  config: Config;
-  isOpen: boolean;
-  onClose?: any;
-}