+++ lib/components/search-modal/components/answer/components/feedback/Feedback.d.ts
@@ -1,6 +1,0 @@
-export type FeedbackProps = {
-  answerId: string;
-  onFeedback: Function;
-  loading?: boolean;
-  success?: boolean;
-}