+++ lib/components/search-modal/components/docs-list/DocsList.d.ts
@@ -1,6 +1,0 @@
-import type { DocType } from '../../services/docs.d';
-
-export type DocsListProps = {
-  docs: Array<DocType>;
-  onClick?: Function;
-}