+++ lib/components/search-modal/components/docs-list/DocsList.tsx
@@ -1,35 +1,67 @@
+import { MouseEvent } from 'react';
 import DocumentIcon from '../../../icons/DocumentIcon';
 import HashIcon from '../../../icons/HashIcon';
+import CloseIcon from '../../../icons/CloseIcon';
+import type { DocType } from '../../services/docs.d';
 import classes from './DocsList.module.css';
-import type { DocsListProps } from './DocsList.d';
 
 const icons = {
   page: DocumentIcon,
   anchor: HashIcon
 };
 
-function DocsList({ docs, onClick }: DocsListProps) {
+type DocsListProps = {
+  docs: Array<DocType>;
+  onClick?: Function;
+  onDelete?: Function;
+}
+
+export default function DocsList ({ docs, onClick, onDelete }: DocsListProps) {
+  function handleClick (doc: DocType) {
+    if (onClick) {
+      onClick(doc);
+    }
+  }
+
+  function handleDelete (event: MouseEvent<HTMLButtonElement>, doc: DocType) {
+    if (onDelete) {
+      event.stopPropagation();
+      event.preventDefault();
+      onDelete(doc);
+    }
+  }
+
   return (
     docs.length ? (
       <ul className={classes.EnhancedSearch__SearchModal__DocsList}>
         {docs.map((doc) => {
           const Icon = (icons[doc.type] || null);
           return (
-            <a key={doc._id} href={doc.url} onClick={onClick ? () => onClick(doc) : undefined}>
+            <a key={doc._id} href={doc.url} onClick={() => handleClick(doc)}>
               <li className={classes.EnhancedSearch__SearchModal__DocsListItem}>
-                {Icon && (
-                  <div className={classes.EnhancedSearch__SearchModal__DocsListItemIcon}>
-                    <Icon />
-                  </div>
-                )}
                 <div>
-                  <span className={classes.EnhancedSearch__SearchModal__DocsListItemTitle}>
-                    {doc.title}
-                  </span>
-                  <span className={classes.EnhancedSearch__SearchModal__DocsListItemDescription}>
-                    {doc.description}
-                  </span>
+                  {Icon && (
+                    <div className={classes.EnhancedSearch__SearchModal__DocsListItemIcon}>
+                      <Icon />
+                    </div>
+                  )}
+                  <div>
+                    <span className={classes.EnhancedSearch__SearchModal__DocsListItemTitle}>
+                      {doc.title}
+                    </span>
+                    <span className={classes.EnhancedSearch__SearchModal__DocsListItemDescription}>
+                      {doc.description}
+                    </span>
+                  </div>
                 </div>
+                {onDelete && (
+                  <button
+                    className={classes.EnhancedSearch__SearchModal__DocsListItemButton}
+                    onClick={(event) => handleDelete(event, doc)}
+                  >
+                    <CloseIcon />
+                  </button>
+                )}
               </li>
             </a>
           )
@@ -38,5 +70,3 @@ function DocsList({ docs, onClick }: DocsListProps) {
     ) : null
   )
 }
-
-export default DocsList;