+++ lib/helpers/local-storage.ts
@@ -1,8 +1,8 @@
-export function getLocalStorageItem(key: string) {
+export function getLocalStorageItem (key: string) {
   const item = localStorage.getItem(key);
   return item ? JSON.parse(item) : null;
 }
 
-export function setLocalStorageItem(key: string, value: any) {
+export function setLocalStorageItem (key: string, value: any) {
   localStorage.setItem(key, JSON.stringify(value));
 }